+++ CHANGELOG.md
@@ -2,3 +2,13 @@
 
 * Implement platform [feishu](https://open.feishu.cn/).
 * Implement target [getx](https://pub.dev/packages/get#translations).
+
+## 0.0.2
+
+* Basic implement for target [arb](https://github.com/google/app-resource-bundle/wiki/ApplicationResourceBundleSpecification).
+
+  **Currently supported:**
+
+  * `description`
+  * Only `String` type `placeholders`
+